moved Encodable  to encoding.rs
