+++ zkvm/src/encoding.rs
@@ -150,3 +150,16 @@ pub(crate) fn write_bytes(x: &[u8], target: &mut Vec<u8>) {
 pub(crate) fn write_point(x: &CompressedRistretto, target: &mut Vec<u8>) {
     write_bytes(x.as_bytes(), target);
 }
+// Encodable provides a set of additional methods to work with bytes
+pub(crate) trait Encodable {
+    ///Encodes receiver into bytes appending them to a provided buffer.
+    fn encode(&self, buf: &mut Vec<u8>);
+    ///Returns precise length in bytes for the serialized representation of the receiver.
+    fn serialized_length(&self) -> usize;
+    /// Encodes the receiver into a newly allocated vector of bytes.
+    fn encode_to_vec(&self) -> Vec<u8> {
+        let mut buf = Vec::with_capacity(self.serialized_length());
+        self.encode(&mut buf);
+        buf
+    }
+}
