+++ zkvm/src/program.rs
@@ -4,6 +4,7 @@ use crate::ops::Instruction;
 use crate::predicate::PredicateTree;
 use crate::scalar_witness::ScalarWitness;
 use crate::types::Data;
+use crate::encoding::Encodable;
 
 use core::borrow::Borrow;
 use merlin::Transcript;
@@ -45,19 +46,6 @@ macro_rules! def_op {
              self
         }
     );
-}
-
-pub(crate) trait Encodable {
-    ///Encodes receiver into bytes appending them to a provided buffer.
-    fn encode(&self, buf: &mut Vec<u8>);
-    ///Returns precise length in bytes for the serialized representation of the receiver.
-    fn serialized_length(&self) -> usize;
-    /// Encodes the receiver into a newly allocated vector of bytes.
-    fn encode_to_vec(&self) -> Vec<u8> {
-        let mut buf = Vec::with_capacity(self.serialized_length());
-        self.encode(&mut buf);
-        buf
-    }
 }
 
 impl Encodable for Program {
