+++ zkvm/src/prover.rs
@@ -8,11 +8,11 @@ use std::collections::VecDeque;
 
 use crate::constraints::Commitment;
 use crate::contract::ContractID;
+use crate::encoding::Encodable;
 use crate::errors::VMError;
 use crate::ops::Instruction;
 use crate::point_ops::PointOp;
 use crate::predicate::Predicate;
-use crate::program::Encodable;
 use crate::program::{Program, ProgramItem};
 use crate::tx::{TxHeader, UnsignedTx};
 use crate::vm::{Delegate, VM};
